+++ notes.md
@@ -13,3 +13,18 @@
 # https://www.fullstackpython.com/flask-code-examples.html    
 
 # https://realpython.com/defining-your-own-python-function/
+
+
+
+
+# Git
+
+# hint:   git config pull.rebase false  # merge (the default strategy)
+# hint:   git config pull.rebase true   # rebase
+# hint:   git config pull.ff only       # fast-forward only
+
+
+# flask
+# pip install -U Flask
+# which updated below
+# Successfully installed Flask-3.0.0 Jinja2-3.1.2 MarkupSafe-2.1.3 Werkzeug-3.0.1 blinker-1.7.0 # click-8.1.7 importlib-metadata-6.8.0 itsdangerous-2.1.2 zipp-3.17.0
